Collect all collections in betty.collections
