+++ betty/collection/keyed.py
