+++ betty/collection/mapping.py
