+++ betty/collection/sequence.py
