+++ betty/collections/keyed/__init__.py
@@ -1,0 +1,3 @@
+"""
+Reusable keyed collections.
+"""
