+++ betty/collections/keyed/adapter.py
