+++ betty/collections/mapping/__init__.py
@@ -1,0 +1,3 @@
+"""
+Reusable mappings.
+"""
