+++ betty/collections/mapping/adapter.py
