+++ betty/collections/plugin/__init__.py
@@ -1,0 +1,3 @@
+"""
+Reusable collections for plugin-related data.
+"""
