+++ betty/collections/sequence/__init__.py
@@ -1,0 +1,3 @@
+"""
+Reusable sequences.
+"""
