+++ betty/collections/sequence/adapter.py
