+++ betty/datas/plugin_manufacturer_sequence.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, final
 
 from betty.collection.sequence import MutableResolvedSequence
-from betty.collection.sequence.adapter import MutableResolvedSequenceAdapter
+from betty.collections.sequence.adapter import MutableResolvedSequenceAdapter
 from betty.datas.aggregate.collection.sequence import SequenceDefinition
 from betty.plugin.cls import PluginClsDefinition
 from betty.plugin.factory import PluginManufacturer, ResolvablePluginManufacturer
